requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.15.2
frozenlist==1.5.0
h11==0.14.0
idna==3.10
kombu==5.5.0
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.2
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2==2.9.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
